Crate foursquare

Foursquare provides a api bindings to the foursquare.com API

Examples

Typical use will require instantiation of a foursquare::Client. This requires a version string, set of foursquare::Credentials and a tokio_core Handle reference.

extern crate foursquare;
extern crate hyper;
extern crate tokio_core;

use tokio_core::reactor::Core;
use foursquare::{Credentials, Client};

fn main() {
  let mut core = Core::new().expect("reactor fail");
  let fs = Client::new(
    "YYYYMMDD",
    Credentials::client(
      "client_id", "client_secret"
    ),
    &core.handle()
  );
}

Access to various services are provided via methods on instances of the Client type.

The convention for executing operations typically looks like client.venues().operation(&OperationOptions) where operation is the name of the operation to perform

Errors

Operations typically result in a foursquare::Future Type which is an alias for the the futures crates Future trait with the Error type fixed to the foursquare::Error type.

Reexports

pub use venue::Venues;

pub use error::Error;

pub use error::ErrorKind;

pub use error::Result;

Modules

error
Client errors

venue
Venue interfaces

Structs

Client
Entry point interface for interacting with Foursquare API

Meta

Response

Enums

Credentials
types of credentials used to authenticate requests

Type Definitions

Future
A type alias for Futures that may return foursquare::Errors