Struct foursquare::venue::ExploreOptionsBuilder

pub struct ExploreOptionsBuilder { /* fields omitted */ }

Builder for ExploreOptions.

Methods

impl ExploreOptionsBuilder

pub fn ll<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

required unless near is provided. Latitude and longitude of the user’s location.

pub fn near<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

required unless ll is provided. A string naming a place in the world. If the near string is not geocodable, returns a failed_geocode error. Otherwise, searches within the bounds of the geocode and adds a geocode object to the response.

pub fn ll_acc<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Accuracy of latitude and longitude, in meters.

pub fn alt<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Altitude of the user’s location, in meters.

pub fn alt_acc<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Accuracy of the user’s altitude, in meters.

pub fn radius<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Radius to search within, in meters. If radius is not specified, a suggested radius will be used based on the density of venues in the area. The maximum supported radius is currently 100,000 meters.

pub fn section<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

One of food, drinks, coffee, shops, arts, outdoors, sights, trending, nextVenues (venues frequently visited after a given venue), or topPicks (a mix of recommendations generated without a query from the user). Choosing one of these limits results to venues with the specified category or property.

pub fn query<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

A term to be searched against a venue’s tips, category, etc. The query parameter has no effect when a section is specified.

pub fn limit<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Number of results to return, up to 50.

pub fn offset<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Used to page through results, up to 50.

pub fn novelty<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

Pass new or old to limit results to places the acting user hasn’t been or has been, respectively. Omitting this parameter returns a mixture of old and new venues.

pub fn friend_visits<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

Pass visited or notvisited to limit results to places the acting user’s friends have or haven’t been, respectively. Omitting this parameter returns a mixture of venues to which the user’s friends have or haven’t been.

pub fn time<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

Pass any to retrieve results for any time of day. Omitting this parameter returns results targeted to the current time of day.

pub fn day<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

Pass any to retrieve results for any day of the week. Omitting this parameter returns results targeted to the current day of the week.

pub fn venue_photos<VALUE: Into<Option<u16>>>(
&mut self,
value: VALUE
) -> &mut Self

Boolean flag to include a photo in the response for each venue, if one is available. Default is 0 (no photos). Photos are returned as part of the venue JSON object.

pub fn last_venue<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

A venue ID to use in combination with the intent=nextVenues parameter, which returns venues users often visit after a given venue. If intent=nextVenues is specified but lastVenue is not, the user’s last check-in will be used if it is within 2 hours. If the user has not checked in within the last 2 hours, no results will be returned.

pub fn open_now<VALUE: Into<Option<u16>>>(&mut self, value: VALUE) -> &mut Self

Boolean flag to only include venues that are open now. This prefers official provider hours but falls back to popular check-in hours.

pub fn sort_by_distance<VALUE: Into<Option<u16>>>(
&mut self,
value: VALUE
) -> &mut Self

Boolean flag to sort the results by distance instead of relevance.

pub fn price<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

Comma separated list of price points. Currently the valid range of price points are [1,2,3,4], 1 being the least expensive, 4 being the most expensive. For food venues, in the United States, 1 is < $10 an entree, 2 is $10-$20 an entree, 3 is $20-$30 an entree, 4 is > $30 an entree.

pub fn saved<VALUE: Into<Option<u16>>>(&mut self, value: VALUE) -> &mut Self

Boolean flag to only include venues that the user has saved on their To-Do list or to another list.

pub fn locale<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

Internationalization

pub fn build(&self) -> Result<ExploreOptions, String>

Builds a new ExploreOptions.

Errors

If a required field has not been initialized.