Struct foursquare::venue::ExploreResponse

pub struct ExploreResponse {
    pub warning: Option<Warning>,
    pub suggested_radius: Option<u32>,
    pub header_location: String,
    pub header_full_location: String,
    pub header_location_granularity: String,
    pub query: Option<String>,
    pub total_results: u64,
    pub groups: Vec<Group<VenueItem>>,
}

Fields
warning: Option<Warning> suggested_radius: Option<u32>

If no radius was specified in the request, presents the radius that was used for the query (based upon the density of venues in the query area).

header_location: String

A text name for the location the user searched, e.g. “SoHo”.

header_full_location: String

A full text name for the location the user searched, e.g. “SoHo, New York”.

header_location_granularity: String query: Option<String> total_results: u64 groups: Vec<Group<VenueItem>>

An array of objects representing groups of recommendations. Each group contains a type such as “recommended” a human-readable (eventually localized) name such as “Recommended Places,” and an array items of recommendation objects, which have an ordered list of objects which contain reasons and venue. The reasons are count and items, where each item has a type such as “social” and a message about why this place may be of interest to the acting user. The venues are compact venues that include stats and hereNow data. We encourage clients to be robust against the introduction or removal of group types by treating the groups as opaque objects to be displayed or by placing unfamiliar groups in a catchall group.