Struct foursquare::venue::RecommendationsOptionsBuilder

pub struct RecommendationsOptionsBuilder { /* fields omitted */ }

Builder for RecommendationsOptions.

Methods

impl RecommendationsOptionsBuilder

pub fn ll<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

required unless near is provided. Latitude and longitude of the user’s location.

pub fn near<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

required unless ll is provided. A string naming a place in the world. If the near string is not geocodable, returns a failed_geocode error. Otherwise, searches within the bounds of the geocode and adds a geocode object to the response.

pub fn ll_acc<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Accuracy of latitude and longitude, in meters.

pub fn alt<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Altitude of the user’s location, in meters.

pub fn alt_acc<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Accuracy of the user’s altitude, in meters.

pub fn radius<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Radius to search within, in meters. If radius is not specified, a suggested radius will be used based on the density of venues in the area. The maximum supported radius is currently 100,000 meters.

pub fn intent<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

One of: food, breakfast, brunch, lunch, coffee, dinner, dessert, drinks, shopping, fun, sights. Specifies the top-level “intent” for a search.

pub fn prices<VALUE: Into<Option<Vec<u16>>>>(
&mut self,
value: VALUE
) -> &mut Self

One of: 1, 2, 3, 4. Only return venues that match the specified price(s), 1 being “$” and 4 being “”. Supports multiple values.

pub fn query<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

A search term to be applied against venue names.

pub fn categories<VALUE: Into<Option<Vec<String>>>>(
&mut self,
value: VALUE
) -> &mut Self

Return values that match the specified categories, after the query parameter is applied.

pub fn category_id<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

Return values that match the specified categories when there is no query parameter provided.

pub fn limit<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Number of results to return, up to 50.

pub fn offset<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Used to page through results, up to 50.

pub fn features<VALUE: Into<Option<Vec<Feature>>>>(
&mut self,
value: VALUE
) -> &mut Self

Specifies what features (takes credit cards, offers delivery, etc.) that the returned venues should have. The following param values correspond to various features this endpoint supports. 0 Takes credit cards. 1 Takes reservations. 2 Offers delivery. 3 Offers take out. 4 Has Wi-Fi. 5 Has outdoor seating. 7 User has liked this venue. 8 Recently opened. 9 Not part of a chain. 10 Takes online reservations. 13 Dog-friendly. 14 Has parking. 15 Has a happy hour.

pub fn open_now<VALUE: Into<Option<bool>>>(&mut self, value: VALUE) -> &mut Self

Boolean flag to only include venues that are open now. This prefers official provider hours but falls back to popular check-in hours.

pub fn sort_by_distance<VALUE: Into<Option<bool>>>(
&mut self,
value: VALUE
) -> &mut Self

Boolean flag to sort the results by distance instead of relevance.

pub fn personalization<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

If you make an authenticated request to this endpoint, you can make results more personal based on the user’s experiences on Swarm and Foursquare (e.g., only return venues that the user has saved to a list or ones that the user has liked before). The following table documents the currently-available personalizations.

pub fn local_day<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

1–7 for Monday–Sunday. Only return results that are open on this day.

pub fn local_time<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

Only return results that are open at this time. HH in 24-hr format.

pub fn locale<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

Internationalization

pub fn build(&self) -> Result<RecommendationsOptions, String>

Builds a new RecommendationsOptions.

Errors

If a required field has not been initialized.