Struct foursquare::venue::SearchOptionsBuilder

pub struct SearchOptionsBuilder { /* fields omitted */ }

Builder for SearchOptions.

Methods

impl SearchOptionsBuilder

pub fn ll<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

required unless near is provided. Latitude and longitude of the user’s location. Optional if using intent=global

pub fn near<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

required unless ll is provided. A string naming a place in the world. If the near string is not geocodable, returns a failed_geocode error. Otherwise, searches within the bounds of the geocode and adds a geocode object to the response.

pub fn intent<VALUE: Into<Option<Intent>>>(&mut self, value: VALUE) -> &mut Self

One of the values below, indicating your intent in performing the search. If no value is specified, defaults to checkin.

pub fn radius<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Limit results to venues within this many meters of the specified location. Defaults to a city-wide area. Only valid for requests with intent=browse, or requests with intent=checkin and categoryId or query. Does not apply to intent=match requests. The maximum supported radius is currently 100,000 meters.

pub fn sw<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

With ne, limits results to the bounding box defined by the latitude and longitude given by sw as its south-west corner, and ne as its north-east corner. The bounding box is only supported for intent=browse searches. Not valid with ll or radius. Bounding boxes with an area up to approximately 10,000 square kilometers are supported.

pub fn ne<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

See sw.

pub fn query<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

A search term to be applied against venue names.

pub fn limit<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Number of results to return, up to 50.

pub fn category_id<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

A comma separated list of categories to limit results to. If you specify categoryId. specifying a radius may improve results. If specifying a top-level category, all sub-categories will also match the query. Does not apply to intent=match requests.

pub fn ll_acc<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Accuracy of latitude and longitude, in meters.

pub fn alt<VALUE: Into<Option<u32>>>(&mut self, value: VALUE) -> &mut Self

Altitude of the user’s location, in meters.

pub fn alt_acc<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Accuracy of the user’s altitude, in meters.

pub fn url<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

A third-party URL which we will attempt to match against our map of venues to URLs.

pub fn provider_id<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

Identifier for a known third party that is part of our map of venues to URLs, used in conjunction with linkedId.

pub fn linked_id<VALUE: Into<Option<String>>>(
&mut self,
value: VALUE
) -> &mut Self

Identifier used by third party specified in providerId, which we will attempt to match against our map of venues to URLs.

pub fn locale<VALUE: Into<Option<String>>>(&mut self, value: VALUE) -> &mut Self

Internationalization

pub fn build(&self) -> Result<SearchOptions, String>

Builds a new SearchOptions.

Errors

If a required field has not been initialized.