Struct foursquare::venue::Venue

pub struct Venue {
    pub id: String,
    pub name: String,
    pub contact: Option<Contact>,
    pub location: Location,
    pub categories: Vec<Category>,
    pub verified: Option<bool>,
    pub url: Option<String>,
    pub hours: Option<Hours>,
    pub has_menu: Option<bool>,
    pub menu: Option<Menu>,
    pub price: Option<Price>,
    pub photos: Option<Photos>,
    pub referral_id: Option<String>,
    pub has_perk: Option<bool>,
    pub rating: Option<f32>,
    pub rating_signals: Option<u64>,
    pub time_zone: Option<String>,
    pub attributes: Option<AttributeGroups>,
    pub canonical_url: Option<String>,
}

Fields
id: String

A unique string identifier for this venue.

name: String

The best known name for this venue.

contact: Option<Contact>

An object containing none, some, or all of twitter, phone, and formattedPhone. All are strings. Will be None for suggest requests

location: Location

An object containing none, some, or all of address (street address), crossStreet, city, state, postalCode, country, lat, lng, and distance. All fields are strings, except for lat, lng, and distance. Distance is measured in meters. Some venues have their locations intentionally hidden for privacy reasons (such as private residences). If this is the case, the parameter isFuzzed will be set to true, and the lat/lng parameters will have reduced precision.

categories: Vec<Category>

An array, possibly empty, of categories that have been applied to this venue. One of the categories will have a primary field indicating that it is the primary category for the venue. For the complete category tree, see categories.

verified: Option<bool>

Boolean indicating whether the owner of this business has claimed it and verified the information. Will be be None for suggest requests

url: Option<String>

URL of the venue’s website, typically provided by the venue manager.

hours: Option<Hours>

Contains the hours during the week that the venue is open along with any named hours segments in a human-readable format. For machine readable hours see venues/hours

has_menu: Option<bool>

Contains the hours during the week when people usually go to the venue. For machine readable hours see venues/hours.

menu: Option<Menu>

An object containing url and mobileUrl that display the menu information for this venue.

price: Option<Price>

An object containing the price tier from 1 (least pricey) - 4 (most pricey) and a message describing the price tier.

photos: Option<Photos>

A count and groups of photos for this venue. Group types are checkin and venue. Not all items will be present. Will typically not be present for search requests

referral_id: Option<String> has_perk: Option<bool> rating: Option<f32>

Numerical rating of the venue (0 through 10). Not all venues will have a rating.

rating_signals: Option<u64> time_zone: Option<String>

time zone, only present in details requests

attributes: Option<AttributeGroups>

Attributes associated with the venue, such as price tier, whether the venue takes reservations, and parking availability. only present in details requests

canonical_url: Option<String>

The canonical URL for this venue, e.g. https://foursquare.com/v/foursquare-hq/4ab7e57cf964a5205f7b20e3 only present in details requests